refactor: move `preProcessRequest()` function under `.../interceptors/helpers`
