+++ packages/hoppscotch-common/src/platform/std/interceptors/proxy.ts
@@ -1,7 +1,7 @@
 import { Interceptor, RequestRunResult } from "~/services/interceptor.service"
 import { AxiosRequestConfig, CancelToken } from "axios"
 import * as E from "fp-ts/Either"
-import { preProcessRequest } from "./browser"
+import { preProcessRequest } from "./helpers"
 import { v4 } from "uuid"
 import axios from "axios"
 import { settingsStore } from "~/newstore/settings"
